requirements.txt
flask==2.3.3
face-recognition==1.3.0
flask-login==0.6.2
werkzeug==2.3.7
pillow==10.0.0
numpy==1.25.2
opencv-python==4.8.0.76
qrcode==7.4.2
python-dotenv==1.0.0
flask-wtf==1.1.1
email-validator==2.0.0.post2